pyproject.toml
[project]
name = "entropy-paper-code"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "scipy",
    "pandas",
    "pyyaml",
    "tqdm",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]